bugfixes and improvements

+++ backend/services/categories.js
@@ -1,4 +1,5 @@
 import Category from '../models/categories.js';
+import Transaction from '../models/transactions.js';
 
 export default class CategoriesService {
   static async getCategories(userId) {
@@ -26,6 +27,10 @@ export default class CategoriesService {
   }
 
   static async removeCategory(categoryId, userId) {
+    const transactionsWithThisCategory = await Transaction.find({
+      categories: categoryId,
+    });
+    if (transactionsWithThisCategory.length !== 0) return null;
     const deletedCategory = await Category.findOneAndDelete({
       _id: categoryId,
       userId,
